- Combine all API keys migrations into a single file - Add idempotent migration script with existence checks - Rename migration to descriptive name (0064_api_keys_user_ownership) - Add helper script for applying migrations safely - Move API keys page from settings to top-level navigation
This migration transforms API keys from project-scoped to user-scoped, allowing them to survive project deletion with ON DELETE SET NULL.
